Cleaning up examples
